server/requirements.txt
paho_mqtt==2.1.0
Pillow==10.3.0
